Commit before Card payment

+++ client/src/pages/Admin/UpdateProduct.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import Layout from "./../../components/Layout/Layout";
 import AdminMenu from "./../../components/Layout/AdminMenu";
-import toast from "react-toastify";
+import {toast} from "react-toastify";
 import axios from "axios";
 import { Select } from "antd";
 import { useNavigate, useParams } from "react-router-dom";
@@ -15,6 +15,7 @@ const UpdateProduct = () => {
   const [description, setDescription] = useState("");
   const [price, setPrice] = useState("");
   const [category, setCategory] = useState("");
+  const [categoryName, setCategoryName] = useState("");
   const [quantity, setQuantity] = useState("");
   const [shipping, setShipping] = useState("");
   const [photo, setPhoto] = useState("");
@@ -34,6 +35,7 @@ const UpdateProduct = () => {
       setQuantity(data.product.quantity);
       setShipping(data.product.shipping);
       setCategory(data.product.category._id);
+      setCategoryName(data.product.category.name);
     } catch (error) {
       console.log(error);
     }
@@ -42,6 +44,7 @@ const UpdateProduct = () => {
     getSingleProduct();
     //eslint-disable-next-line
   }, []);
+  
   //get all category
   const getAllCategory = async () => {
     try {
@@ -54,10 +57,26 @@ const UpdateProduct = () => {
       toast.error("Something wwent wrong in getting catgeory");
     }
   };
-
   useEffect(() => {
     getAllCategory();
   }, []);
+  const option=[];
+  categories?.map((c) => (
+    option.push({
+      value:c.name,
+      label:c.name,
+      
+    }) 
+  ))
+  const handleChange= (e)=> {
+    {categories?.map((c) => {
+      if(c.name===e){
+        setCategory(c._id);
+        setCategoryName(c.name)
+      }
+    }) } 
+  }
+  
 
   //create product function
   const handleUpdate = async (e) => {
@@ -70,15 +89,16 @@ const UpdateProduct = () => {
       productData.append("quantity", quantity);
       photo && productData.append("photo", photo);
       productData.append("category", category);
-      const { data } = axios.put(
+      const { data } = await axios.put(
         `/api/v1/product/update-product/${id}`,
         productData
       );
-      if (data?.success) {
-        toast.error(data?.message);
-      } else {
+      if (data.success) {
         toast.success("Product Updated Successfully");
         navigate("/dashboard/admin/products");
+        
+      } else {
+        toast.error(data?.message);
       }
     } catch (error) {
       console.log(error);
@@ -90,11 +110,12 @@ const UpdateProduct = () => {
   const handleDelete = async () => {
     try {
       let answer = window.prompt("Are You Sure want to delete this product ? ");
-      if (!answer) return;
+      if (!answer || (answer!=="yes") ) return;
+      // console.log("hi");
       const { data } = await axios.delete(
         `/api/v1/product/delete-product/${id}`
       );
-      toast.success("Product DEleted Succfully");
+      toast.success("Product Deleted Succesfully");
       navigate("/dashboard/admin/products");
     } catch (error) {
       console.log(error);
@@ -111,23 +132,22 @@ const UpdateProduct = () => {
           <div className="col-md-9">
             <h1>Update Product</h1>
             <div className="m-1 w-75">
-              <Select
-                bordered={false}
-                placeholder="Select a category"
-                size="large"
+            <Select
+                // bordered={false}
+                // mode="tags"
                 showSearch
-                className="form-select mb-3"
-                onChange={(value) => {
-                  setCategory(value);
+                placeholder="Select a category"
+                style={{
+                  width: '100%',
                 }}
-                value={category}
-              >
-                {categories?.map((c) => (
-                  <Option key={c._id} value={c._id}>
-                    {c.name}
-                  </Option>
-                ))}
-              </Select>
+                // size="large"
+                // value={category}
+                className=" mb-3"
+                
+                onChange={handleChange}
+                value={categoryName}
+                options={option}
+               />
               <div className="mb-3">
                 <label className="btn btn-outline-secondary col-md-12">
                   {photo ? photo.name : "Upload Photo"}
@@ -199,20 +219,25 @@ const UpdateProduct = () => {
                 />
               </div>
               <div className="mb-3">
-                <Select
-                  bordered={false}
+              <Select
                   placeholder="Select Shipping "
-                  size="large"
-                  showSearch
-                  className="form-select mb-3"
+                  // showSearch
+                  className="mb-3"
                   onChange={(value) => {
                     setShipping(value);
+                    // console.log(value);
                   }}
-                  value={shipping ? "yes" : "No"}
-                >
-                  <Option value="0">No</Option>
-                  <Option value="1">Yes</Option>
-                </Select>
+                  value={shipping ? "Yes" : "No"}
+                  options={[{
+                    value:0,
+                    label:"No"
+                  },
+                  {
+                    value:1,
+                    label:"Yes"
+                  }
+                  ]}
+                />
               </div>
               <div className="mb-3">
                 <button className="btn btn-primary" onClick={handleUpdate}>
